# example_trello_swagger.yaml
openapi: 3.0.0
info:
  title: OpenAPI спецификация example_trello
  version: 0.0.1
servers:
  - url: http://localhost:8080/api/
    description: Dev server  
paths:
...
  /send-pincode:
    post:
      summary: Метод отправляет введенный пин-код пользователем на сервер.
      description: Метод предназначен для подтверждения личности данного пользователя и сопровождения на финальный этап сброса пароля
      operationId: sendPincode
      tags:
        - Reset Password
      requestBody:
        required: true
        content:
          'application/json':
            schema:
              $ref: "#/components/schemas/SendPincode"
      responses:
        '200':
          "description": Подтверждение успешной отправки
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SendPincode"
        '400':
          "description": Некорректные входные данные. Возвращает список атрибутов с ошибками
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
        default:
          description: Любая неожиданная ошибка
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /reset-password:
    post:
      summary: Метод отправляет пароль и пароль-подтверждение на сервер.
      description: Метод предназначен для перезаписи старого пароля на новый.
      operationId: resetPassword
      tags:
        - Reset Password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPassword"
      responses:
        '200':
          description: Подтверждение смены пароля
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResetPassword"
        '400':
          description: Некорректные входные данные. Возвращает список атрибутов с ошибками
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
        default:
          description: Любая неожиданная ошибка
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  schemas:
...
    ResetPassword:
      type: object
      required:
        - newPassword
        - confirmPassword
      properties:
        newPassword:
          type: string
          example: "klimushkin1988"
        confirmPassword:
          type: string
          example: "klimushkin1988"
    SendPincode:
      type: object
      required:
        - email
        - pincode
      properties:
        email:
          type: string
          example: muhammed1942ali@gmail.com
        pincode:
          type: string
          example: 13bbgh
...
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string